tests: add SIG(,) to test repertoire
allow test sigs to specified using SIG(privkey,hash).

if SIG() is specified in a `expect-send` event, test runner will verify
that the node's signature is valid for the given key and hash.

if SIG() is specified in a `recv` event, test runner will generate
a valid signature of the hash with the provided privkey and send
that to the peer.