df: set max number of messages for each of the tx collabs
To prevent infinite loops, we set the max additions/subtractions to
2^12 (4,096).

If each side sends the maximum allowed messages, this is a total of
32,768 messages (2^12 * 4 * 2). (The end result of this exchange would
necessarily be an empty tx, as the number of adds would equal the number
of removals).

In practice, this bounds the total inputs/outputs to 2*13 (8,192) each.